Improve performance
Replace regex parsing with proper java source file parser
Organize code a bit better, with Grouper class

This version doesn't have the false-positive detection on well-formed
Java files, but I have noticed that Java files which put the package
statement before the compilation unit / file header comment do get that
comment duplicated.

This version handles multiple imports on a single line perfectly well.

This version does have issues if the first import occurs on the same
line as the package statement, or if the last import occurs on the same
line as the beginning of the class. This should be corrected in a future
version (See issue #1).