Add an option to remove unused imports (#6)
Fixes #5

The algorithm for calculating whether an import is unused, is really
simple. It only looks at the body of the file, and checks if the last
segment of the import path (typically a class name or a static function
name) is refernenced in the code. This makes us unable to remove imports
with wildcards, but apart from that, it seems to work well.

The "removeUnused" configuration parameter is false by default.

Bump JavaParser version to 3.5.0.

... and start using TokenRange.iterator()