Refactor the code to avoid boolean parameter to ImpSort
This changes the architecture of impsort to have two separate components
called the Parser and the Grouper. The Parser parses the file and
creates a Parser.Result object. This Parser.Result object has a function
that can be called to remove unused imports.

The Parser.Result object can be passed into the Grouper, who groups the
imports according to the rules, and creates a Grouper.Result object. This
Grouper.Result object contains the sorted and grouped imports, and may be
saved to disk.