Set surefire's tempdir to target
This patch follows up on the review given for #51 and uses
maven-surefire-plugin's configuration to set the tempdir to target and
restore the behavior change introduced in that PR.

Note: This fix will only effect running the test via maven. If the
test is manually run (e.g., from an IDE), java.io.tmpdir should be
overridden manually too.